PHPStan fixes

+++ api/src/AppBundle/Service/ReportService.php
@@ -20,14 +20,18 @@ use AppBundle\Entity\Report\Document;
 use AppBundle\Entity\Report\Report;
 use AppBundle\Entity\Report\ReportSubmission;
 use AppBundle\Entity\ReportInterface;
+use AppBundle\Entity\Repository\ReportRepository;
 use AppBundle\Entity\User;
+use DateTime;
 use Doctrine\Common\Collections\Collection;
+use Doctrine\Common\Persistence\ObjectRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\EntityRepository;
+use Exception;
 
 class ReportService
 {
-    /** @var EntityRepository */
+    /** @var ReportRepository */
     protected $reportRepository;
 
     /**
@@ -36,42 +40,38 @@ class ReportService
     protected $_em;
 
     /**
-     * @var EntityRepository
+     * @var ObjectRepository
      */
     private $casRecRepository;
 
     /**
-     * @var EntityRepository
+     * @var ObjectRepository
      */
     private $assetRepository;
 
     /**
-     * @var EntityRepository
+     * @var ObjectRepository
      */
     private $bankAccountRepository;
 
     public function __construct(
-        EntityManagerInterface $em
+        EntityManagerInterface $em,
+        ReportRepository $reportRepository
     )
     {
-        $this->reportRepository = $em->getRepository(Report::class);
+        $this->reportRepository = $reportRepository;
         $this->casRecRepository = $em->getRepository(CasRec::class);
         $this->_em = $em;
         $this->assetRepository = $em->getRepository(Asset::class);
         $this->bankAccountRepository = $em->getRepository(BankAccountEntity::class);
     }
 
-    public function findById($id)
-    {
-        return $this->reportRepository->findOneBy(['id' => $id]);
-    }
-
     /**
      * Set report type based on CasRec record (if existing)
      *
-     * @param Client $report
-     *
+     * @param Client $client
      * @return string|null type
+     * @throws Exception
      */
     public function getReportTypeBasedOnCasrec(Client $client)
     {
@@ -86,9 +86,10 @@ class ReportService
     /**
      * Create new year's report copying data over (and set start/endDate accordingly).
      *
-     * @param Report $oldReport
+     * @param ReportInterface $oldReport
      *
      * @return Report
+     * @throws Exception
      */
     private function createNextYearReport(ReportInterface $oldReport)
     {
@@ -102,11 +103,14 @@ class ReportService
             $startDate = clone $oldReport->getEndDate();
             $newReportType = $this->getReportTypeBasedOnCasrec($client) ?: $oldReport->getType();
             $startDate->modify('+1 day');
-        } else {
+        } elseif ($oldReport instanceof Ndr) {
             // when the previous report is NDR we need to work out the new reporting period
+            /** @var DateTime $startDate */
             $startDate = $oldReport->getClient()->getExpectedReportStartDate();
             // set default type as oldReport is ndr
             $newReportType = $this->getReportTypeBasedOnCasrec($client) ?: Report::TYPE_102;
+        } else {
+            throw new \RuntimeException('createNextYearReport() only supports Report and Ndr');
         }
 
         $endDate = clone $startDate;
@@ -134,7 +138,7 @@ class ReportService
      * @param Report $toReport
      * @param Ndr|Report $fromReport
      */
-    public function clonePersistentResources($toReport, $fromReport)
+    public function clonePersistentResources(Report $toReport, $fromReport)
     {
         // copy assets
         $toReport->setNoAssetToAdd($fromReport->getNoAssetToAdd());
@@ -144,9 +148,11 @@ class ReportService
             // Check that the target report doesn't already have a matching asset
             $assetExists = $this->checkAssetExists($toReport, $asset);
 
-            if (!$assetExists) {
+            if (!$assetExists && $asset instanceof Asset) {
+                /** @var Asset $newAsset */
                 $newAsset = $this->cloneAsset($asset);
                 $newAsset->setReport($toReport);
+
                 $toReport->addAsset($newAsset);
                 $this->_em->detach($newAsset);
                 $this->_em->persist($newAsset);
@@ -169,7 +175,7 @@ class ReportService
 
     /**
      * @param ReportInterface $toReport
-     * @param $asset
+     * @param AssetInterface $asset
      * @return bool
      */
     private function checkAssetExists(ReportInterface $toReport, AssetInterface $asset)
@@ -267,7 +273,7 @@ class ReportService
         $newAccount->setAccountNumber($account->getAccountNumber());
         $newAccount->setOpeningBalance($account->getClosingBalance());
         $newAccount->setIsJointAccount($account->getIsJointAccount());
-        $newAccount->setCreatedAt(new \DateTime());
+        $newAccount->setCreatedAt(new DateTime());
 
         return $newAccount;
     }
@@ -277,12 +283,12 @@ class ReportService
      *
      * @param ReportInterface $currentReport
      * @param User $user
-     * @param \DateTime $submitDate
+     * @param DateTime $submitDate
      * @param null $ndrDocumentId
      * @return Report
      *
      */
-    public function submit(ReportInterface $currentReport, User $user, \DateTime $submitDate, $ndrDocumentId = null)
+    public function submit(ReportInterface $currentReport, User $user, DateTime $submitDate, $ndrDocumentId = null)
     {
         if (!$currentReport->getAgreedBehalfDeputy()) {
             throw new \RuntimeException('Report must be agreed for submission');
@@ -334,11 +340,11 @@ class ReportService
 
     /**
      * @param Report $report
-     * @param \DateTime $unsubmitDate
-     * @param \DateTime $dueDate
+     * @param DateTime $unsubmitDate
+     * @param DateTime $dueDate
      * @param $sectionList
      */
-    public function unSubmit(Report $report, \DateTime $unsubmitDate, \DateTime $dueDate, \DateTime $startDate, \DateTime $endDate, $sectionList)
+    public function unSubmit(Report $report, DateTime $unsubmitDate, DateTime $dueDate, DateTime $startDate, DateTime $endDate, $sectionList)
     {
         // reset report.submitted so that the deputy will set the report back into the dashboard
         $report->setSubmitted(false);
@@ -358,11 +364,11 @@ class ReportService
      *
      * @param Report $currentReport
      * @param User $user
-     * @param \DateTime $submitDate
+     * @param DateTime $submitDate
      *
      * @return Report new year's report
      */
-    public function submitAdditionalDocuments(Report $currentReport, User $user, \DateTime $submitDate)
+    public function submitAdditionalDocuments(Report $currentReport, User $user, DateTime $submitDate)
     {
         // create submission record with NEW documents (= documents not yet attached to a submission)
         $submission = new ReportSubmission($currentReport, $user);
@@ -427,11 +433,11 @@ class ReportService
      * In all the the cases, return original $status
      *
      * @param $status
-     * @param \DateTime $endDate
+     * @param DateTime $endDate
      *
      * @return string
      */
-    public function adjustReportStatus($status, \DateTime $endDate)
+    public function adjustReportStatus($status, DateTime $endDate)
     {
         if ($status == Report::STATUS_READY_TO_SUBMIT && !self::isDue($endDate)) {
             return Report::STATUS_NOT_FINISHED;
@@ -442,16 +448,16 @@ class ReportService
 
 
     /**
-     * @param \DateTime|null $endDate
+     * @param DateTime|null $endDate
      * @return bool
      */
-    public static function isDue(\DateTime $endDate = null)
+    public static function isDue(DateTime $endDate = null)
     {
         if (!$endDate) {
             return false;
         }
 
-        $endOfToday = new \DateTime('today midnight');
+        $endOfToday = new DateTime('today midnight');
 
         return $endDate <= $endOfToday;
     }
